requirements.txt
Babel==2.7.0
Django==2.1.8
django-phone-field==1.8.0
django-phonenumber-field==3.0.1
django-widget-tweaks==1.4.5
phonenumbers==8.10.23
pytz==2019.3
